# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmaSafe Intelligence API",
    "description": "AI-Powered Pharmaceutical Reverse Chain & Medicine Safety Platform Backend Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/intelligence/risk-score/{batch_id_or_number}": {
      "get": {
        "tags": [
          "AI & Risk Intelligence"
        ],
        "summary": "Get Batch Risk Score",
        "operationId": "get_batch_risk_score_api_v1_intelligence_risk_score__batch_id_or_number__get",
        "parameters": [
          {
            "name": "batch_id_or_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id Or Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/online-surveillance": {
      "post": {
        "tags": [
          "AI & Risk Intelligence"
        ],
        "summary": "Ingest Online Listing",
        "operationId": "ingest_online_listing_api_v1_intelligence_online_surveillance_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnlineSurveillanceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineSurveillanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "AI & Risk Intelligence"
        ],
        "summary": "List Online Surveillance",
        "operationId": "list_online_surveillance_api_v1_intelligence_online_surveillance_get",
        "parameters": [
          {
            "name": "high_risk_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "High Risk Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OnlineSurveillanceResponse"
                  },
                  "title": "Response List Online Surveillance Api V1 Intelligence Online Surveillance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/anomalies": {
      "get": {
        "tags": [
          "AI & Risk Intelligence"
        ],
        "summary": "List Anomalies",
        "operationId": "list_anomalies_api_v1_intelligence_anomalies_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnomalyResponse"
                  },
                  "title": "Response List Anomalies Api V1 Intelligence Anomalies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnlineSurveillanceCreate": {
        "properties": {
          "platform_name": {
            "type": "string",
            "title": "Platform Name"
          },
          "listing_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Url"
          },
          "seller_name": {
            "type": "string",
            "title": "Seller Name"
          },
          "seller_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seller Identifier"
          },
          "medicine_brand_claimed": {
            "type": "string",
            "title": "Medicine Brand Claimed"
          },
          "extracted_batch_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Batch Number"
          },
          "listed_price_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listed Price Inr"
          },
          "discount_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Percentage"
          }
        },
        "type": "object",
        "required": [
          "platform_name",
          "seller_name",
          "medicine_brand_claimed"
        ],
        "title": "OnlineSurveillanceCreate"
      },
      "OnlineSurveillanceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "platform_name": {
            "type": "string",
            "title": "Platform Name"
          },
          "listing_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Url"
          },
          "seller_name": {
            "type": "string",
            "title": "Seller Name"
          },
          "medicine_brand_claimed": {
            "type": "string",
            "title": "Medicine Brand Claimed"
          },
          "extracted_batch_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Batch Number"
          },
          "listed_price_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listed Price Inr"
          },
          "discount_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Percentage"
          },
          "is_dead_batch_match": {
            "type": "boolean",
            "title": "Is Dead Batch Match"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevelEnum"
          },
          "flagged_reasons": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flagged Reasons"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "platform_name",
          "seller_name",
          "medicine_brand_claimed",
          "is_dead_batch_match",
          "risk_score",
          "risk_level",
          "detected_at"
        ],
        "title": "OnlineSurveillanceResponse"
      },
...
      "RiskLevelEnum": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "RiskLevelEnum"
      },
      "RiskScoreResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "composite_risk_score": {
            "type": "number",
            "title": "Composite Risk Score"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevelEnum"
          },
          "expiry_risk_score": {
            "type": "number",
            "title": "Expiry Risk Score"
          },
          "supply_chain_anomaly_score": {
            "type": "number",
            "title": "Supply Chain Anomaly Score"
          },
          "reentry_risk_score": {
            "type": "number",
            "title": "Reentry Risk Score"
          },
          "seller_reputation_score": {
            "type": "number",
            "title": "Seller Reputation Score"
          },
          "explanation_summary": {
            "type": "string",
            "title": "Explanation Summary"
          },
          "factor_breakdown": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factor Breakdown"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "composite_risk_score",
          "risk_level",
          "expiry_risk_score",
          "supply_chain_anomaly_score",
          "reentry_risk_score",
          "seller_reputation_score",
          "explanation_summary",
          "timestamp"
        ],
        "title": "RiskScoreResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}